--- lattice/samples/grpc-proto3-lattice-sample/src/main/java/com/flipkart/krystal/lattice/samples/grpc/proto/sampleProtoService/app/CustomInjectionSample.java
@@ -1,3 +1,0 @@
-package com.flipkart.krystal.lattice.samples.grpc.proto.sampleProtoService.app;
-
-public interface CustomInjectionSample {}
