+++ lattice/samples/grpc-proto3-lattice-sample/src/main/java/com/flipkart/krystal/lattice/samples/grpc/proto3/sampleProtoService/SubMessage.java
@@ -1,4 +1,4 @@
-package com.flipkart.krystal.lattice.samples.grpc.proto.sampleProtoService;
+package com.flipkart.krystal.lattice.samples.grpc.proto3.sampleProtoService;
 
 import static com.flipkart.krystal.model.IfAbsent.IfAbsentThen.ASSUME_DEFAULT_VALUE;
 import static com.flipkart.krystal.model.IfAbsent.IfAbsentThen.FAIL;
