+++ lattice/samples/grpc-proto3-lattice-sample/src/main/java/com/flipkart/krystal/lattice/samples/grpc/proto3/sampleProtoService/app/CustomGuiceModule.java
@@ -1,4 +1,4 @@
-package com.flipkart.krystal.lattice.samples.grpc.proto.sampleProtoService.app;
+package com.flipkart.krystal.lattice.samples.grpc.proto3.sampleProtoService.app;
 
 import com.flipkart.krystal.krystex.kryon.KryonExecutorConfigurator;
 import com.google.inject.AbstractModule;
