+++ lattice/samples/grpc-proto3-lattice-sample/src/main/java/com/flipkart/krystal/lattice/samples/grpc/proto3/sampleProtoService/app/CustomInjectionSample.java
@@ -1,0 +1,3 @@
+package com.flipkart.krystal.lattice.samples.grpc.proto3.sampleProtoService.app;
+
+public interface CustomInjectionSample {}
