+++ lattice/samples/grpc-proto3-lattice-sample/src/test/java/com/flipkart/krystal/lattice/samples/grpc/proto3/Proto3LatticeSampleTest.java
@@ -1,4 +1,4 @@
-package com.flipkart.krystal.lattice.samples.proto3;
+package com.flipkart.krystal.lattice.samples.grpc.proto3;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -11,17 +11,17 @@ import com.flipkart.krystal.data.Errable;
 import com.flipkart.krystal.krystex.caching.TestRequestLevelCache;
 import com.flipkart.krystal.krystex.kryon.KryonExecutionConfig;
 import com.flipkart.krystal.krystex.kryon.KryonExecutorConfig;
-import com.flipkart.krystal.lattice.samples.grpc.proto.sampleProtoService.Proto3LatticeSample;
-import com.flipkart.krystal.lattice.samples.grpc.proto.sampleProtoService.Proto3LatticeSampleResponse;
-import com.flipkart.krystal.lattice.samples.grpc.proto.sampleProtoService.Proto3LatticeSampleResponse_Immut;
-import com.flipkart.krystal.lattice.samples.grpc.proto.sampleProtoService.Proto3LatticeSampleResponse_ImmutPojo;
-import com.flipkart.krystal.lattice.samples.grpc.proto.sampleProtoService.Proto3LatticeSampleResponse_ImmutProto3;
-import com.flipkart.krystal.lattice.samples.grpc.proto.sampleProtoService.Proto3LatticeSample_Req;
-import com.flipkart.krystal.lattice.samples.grpc.proto.sampleProtoService.Proto3LatticeSample_ReqImmut;
-import com.flipkart.krystal.lattice.samples.grpc.proto.sampleProtoService.Proto3LatticeSample_ReqImmutProto3;
-import com.flipkart.krystal.lattice.samples.grpc.proto.sampleProtoService.ProtoMessageProto3;
-import com.flipkart.krystal.lattice.samples.grpc.proto.sampleProtoService.ProtoMessage_ImmutProto3;
-import com.flipkart.krystal.lattice.samples.grpc.proto.sampleProtoService.Status;
+import com.flipkart.krystal.lattice.samples.grpc.proto3.sampleProtoService.Proto3LatticeSample;
+import com.flipkart.krystal.lattice.samples.grpc.proto3.sampleProtoService.Proto3LatticeSampleResponse;
+import com.flipkart.krystal.lattice.samples.grpc.proto3.sampleProtoService.Proto3LatticeSampleResponse_Immut;
+import com.flipkart.krystal.lattice.samples.grpc.proto3.sampleProtoService.Proto3LatticeSampleResponse_ImmutPojo;
+import com.flipkart.krystal.lattice.samples.grpc.proto3.sampleProtoService.Proto3LatticeSampleResponse_ImmutProto3;
+import com.flipkart.krystal.lattice.samples.grpc.proto3.sampleProtoService.Proto3LatticeSample_Req;
+import com.flipkart.krystal.lattice.samples.grpc.proto3.sampleProtoService.Proto3LatticeSample_ReqImmut;
+import com.flipkart.krystal.lattice.samples.grpc.proto3.sampleProtoService.Proto3LatticeSample_ReqImmutProto3;
+import com.flipkart.krystal.lattice.samples.grpc.proto3.sampleProtoService.ProtoMessageProto3;
+import com.flipkart.krystal.lattice.samples.grpc.proto3.sampleProtoService.ProtoMessage_ImmutProto3;
+import com.flipkart.krystal.lattice.samples.grpc.proto3.sampleProtoService.Status;
 import com.flipkart.krystal.pooling.Lease;
 import com.flipkart.krystal.pooling.LeaseUnavailableException;
 import com.flipkart.krystal.vajram.VajramDef;
