+++ lattice/samples/grpc-proto3-lattice-sample/src/test/java/com/flipkart/krystal/lattice/samples/grpc/proto3/sampleProtoService/ProtoListBuilderTest.java
@@ -1,4 +1,4 @@
-package com.flipkart.krystal.lattice.samples.grpc.proto.sampleProtoService;
+package com.flipkart.krystal.lattice.samples.grpc.proto3.sampleProtoService;
 
 import static org.assertj.core.api.Assertions.assertThat;
