+++ lattice/samples/rest/json/dropwizard/rest-json-dropwizard-lattice-sample/build.gradle
@@ -55,7 +55,6 @@ extraJavaModuleInfo {
 }
 
 test {
-    useJUnitPlatform()
     // Fork a fresh JVM per test class so the embedded Dropwizard app booted by
     // RestEndpointsE2eTest doesn't share Guice/Jetty state with other suites.
     forkEvery = 1
